moved nbt files into nbt folder
